+++ include/sgl_kernel_ops.h
@@ -496,22 +496,13 @@ void hc_post(
  */
 void hc_pre_gemm_sqr_sum(at::Tensor& C, at::Tensor& sqr_sum, const at::Tensor& A, const at::Tensor& B);
 
-std::tuple<at::Tensor, at::Tensor, at::Tensor, at::Tensor> mhc_fused_post_pre(
+std::tuple<at::Tensor, at::Tensor, at::Tensor> mhc_fused_post_pre_fma(
     const at::Tensor& x,
     const at::Tensor& residual,
     const at::Tensor& post_layer_mix,
     const at::Tensor& comb_res_mix,
     const at::Tensor& fn,
-    const at::Tensor& hc_scale,
-    const at::Tensor& hc_base,
-    double rms_eps,
-    double hc_pre_eps,
-    double hc_sinkhorn_eps,
-    double hc_post_mult_value,
-    int64_t sinkhorn_repeat,
-    int64_t n_splits,
-    std::optional<at::Tensor> norm_weight = std::nullopt,
-    std::optional<double> norm_eps = std::nullopt);
+    int64_t n_splits = 0);
 
 /*
  * From csrc/speculative
